feat: task-runner (Phase 1): Implement procedural name generation for seed-derived materials (extend/extract fabricator's `procedural_name` into a shared location — both fabricator and material derivation need it)
